feat!: changed the style of consultationReport and prescription interfaces, added mails files for sending mails to patient when doctor change the status of an appointment(confirmed, refused, cancelled or updated)

+++ app/Http/Controllers/ExamResultController.php
@@ -2,11 +2,11 @@
 
 namespace App\Http\Controllers;
 
-use App\Models\ConsultationReport;
 use App\Models\ExamResult;
 use App\Models\MedicalRecord;
 use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Storage;
 
 class ExamResultController extends Controller
 {
@@ -16,6 +16,7 @@ class ExamResultController extends Controller
     public function index($patient_id)
     {
         $patient = User::findById($patient_id);
+        $this->authorize('view', $patient->medicalRecord);
         $examResults = ExamResult::whereHas('medical_record', function ($query) use ($patient) {
             $query->where('patient_id', $patient->id);
         })->get();
@@ -25,33 +26,45 @@ class ExamResultController extends Controller
     /**
      * Show the form for creating a new resource.
      */
-    public function create()
+    public function create($medicalRecord_id)
     {
+        $medicalRecord = MedicalRecord::findOrFail($medicalRecord_id);
+        $patient = $medicalRecord->patient;
+
+        $this->authorize('create', [ExamResult::class, $patient->id]);
+
         return view('exam_results.create');
     }
 
     /**
      * Store a newly created resource in storage.
      */
-    public function store(Request $request, $patient_id)
+    public function store(Request $request, $medicalRecord_id)
     {
-        $request->validate([
-            'type' => 'required|string|max:255',
+        $medicalRecord = MedicalRecord::findOrFail($medicalRecord_id);
+        $patient = $medicalRecord->patient;
+
+        $this->authorize('create', [ExamResult::class, $patient->id]);
+
+        $validatedData = $request->validate([
+            'type'   => 'required|string|max:255',
             'result' => 'nullable|string',
-            'doc' => 'nullable',
-            'date' => 'nullable|date'
+            'doc'    => 'nullable|file|mimes:jpeg,png,jpg,pdf,doc,docx|max:2048',
+            'date'   => 'nullable|date',
         ]);
 
-        $medicalRecord = MedicalRecord::where('patient_id', $patient_id)->first();
-        $medicalRecord_id = $medicalRecord->id;
-        ExamResult::create([
-            'medicalRecord_id' => $medicalRecord_id,
-            'type' => $request->type,
-            'result' => $request->result,
-            'date' => $request->date
-        ]);
+        if ($request->hasFile('doc')) {
+            $file = $request->file('doc');
+            $filename = time() . '_' . $file->getClientOriginalName();
+            $filepath = $file->storeAs('exam_results', $filename, 'public');
+            $validatedData['doc'] = $filepath;
+        }
 
-        return redirect()->route('exam_results.index')->with('success', 'Exam result added successfully.');
+        $validatedData['medicalRecord_id'] = $medicalRecord_id;
+        ExamResult::create($validatedData);
+
+        return redirect()->back();
+//        return redirect()->route('exam_results.index')->with('success', 'Exam result added successfully.');
     }
 
     /**
@@ -73,31 +86,59 @@ class ExamResultController extends Controller
     /**
      * Update the specified resource in storage.
      */
-    public function update(Request $request, ExamResult $examResult)
+    public function update(Request $request, $examResult_id)
     {
-        $request->validate([
-            'type' => 'required|string|max:255',
+        $examResult = ExamResult::findOrFail($examResult_id);
+        $this->authorize('update', $examResult);
+
+        $validatedData = $request->validate([
+            'type'   => 'required|string|max:255',
             'result' => 'nullable|string',
-            'doc' => 'nullable|file|mimes:pdf,jpeg,png,jpg,gif,svg,csv',
-            'date' => 'nullable|date'
+            'doc'    => 'nullable|file|mimes:jpeg,png,jpg,pdf,doc,docx|max:2048',
+            'date'   => 'nullable|date',
         ]);
 
-        $examResult::update([
-            'type' => $request->type,
-            'value' => $request->value,
-            'unit' => $request->unit
-        ]);
+        $examResult->update($validatedData);
 
-        return redirect()->route('exam_results.index')->with('success', 'Exam result updated successfully.');
+        if ($request->hasFile('doc')) {
+
+            if ($examResult->doc) {
+                Storage::disk('public')->delete($examResult->doc);
+            }
+            $file = $request->file('doc');
+            $filename = time() . '_' . $file->getClientOriginalName();
+            $path = $file->storeAs('exam_results', $filename, 'public');
+            $examResult->doc = $path;
+        }
+        $examResult->save();
+
+        return redirect()->back();
+//        return redirect()->route('exam_results.index')->with('success', 'Exam result updated successfully.');
     }
 
+    public function downloadDoc($doc)
+    {
+        // Retrieve the path to the document
+        $filePath = 'exam_results/' . $doc;
+
+        if (!Storage::disk('public')->exists($filePath)) {
+            return redirect()->back()->with('error', 'Le document n\'existe pas.');
+        }
+
+//        return Storage::download($filePath);
+        return response()->download($filePath);
+    }
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(ExamResult $examResult)
+    public function destroy($examResult_id)
     {
+        $examResult = ExamResult::findOrFail($examResult_id);
+        $this->authorize('delete', $examResult);
+
         $examResult->delete();
 
-        return redirect()->route('exam_results.index')->with('success', 'Exam result deleted successfully.');
+        return redirect()->back();
+//        return redirect()->route('exam_results.index')->with('success', 'Exam result deleted successfully.');
     }
 }
